Create GCD using Euclidian Algorithm
The Euclidean algorithm is an efficient method for finding the greatest common divisor (GCD) of two integers. It works by repeatedly subtracting the smaller integer from the larger integer until the remainder is zero. The GCD of the two integers is the last non-zero remainder.
For example, to find the GCD of 15 and 20, we first subtract 15 from 20 to get a remainder of 5. Then we subtract 5 from 15 to get a remainder of 0. Therefore, the GCD of 15 and 20 is 5.
The Euclidean algorithm can be generalized to find the GCD of any two integers. The steps are as follows:
1. Start with two integers, a and b.
2. If b is equal to 0, then the GCD is a.
3. Otherwise, subtract b from a to get a remainder r.
4. Replace a with b and b with r.
5. Go back to step 2.
The GCD of two integers is always the smallest positive integer that divides both integers evenly. The Euclidean algorithm can be used to find the GCD of any two integers, even if they are not coprime (that is, they do not have a common factor other than 1).