Bubble Sort
Bubble sort is a sorting algorithm that works by repeatedly comparing adjacent elements in an array and swapping them if they are in the wrong order. The algorithm continues to do this until the array is sorted.
Bubble sort is a simple algorithm, but it is also inefficient. It has a time complexity of O(n^2), which means that it takes quadratic time to sort an array of size n. This makes bubble sort impractical for large arrays.
Despite its inefficiency, bubble sort is still a useful algorithm to know. It is a simple algorithm that is easy to understand and implement. Bubble sort can also be used as a building block for more efficient sorting algorithms.